% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocar_municipios.R
\name{alocar_municipios}
\alias{alocar_municipios}
\title{Alocação Otimizada de Unidades de Coleta (UCs) a Agências}
\usage{
alocar_municipios(
  ucs,
  agencias = data.frame(agencia_codigo = unique(ucs$agencia_codigo), uc_agencia_max =
    Inf, custo_fixo = 0),
  custo_litro_combustivel = 6,
  custo_hora_viagem = 10,
  kml = 10,
  valor_diaria = 335,
  diarias_entrevistador_max = Inf,
  remuneracao_entrevistador = 0,
  n_entrevistadores_min = 1,
  dias_coleta_entrevistador_max,
  dias_treinamento = 0,
  agencias_treinadas = NULL,
  agencias_treinamento = NULL,
  distancias_ucs,
  distancias_agencias = NULL,
  uc_agencia_min = 1,
  adicional_troca_jurisdicao = 0,
  resultado_completo = FALSE,
  solver = "cbc",
  ...
)
}
\arguments{
\item{ucs}{Um \code{tibble} ou \code{data.frame} contendo informações sobre as UCs, incluindo:
\itemize{
\item \code{uc}: Código único da UC.
\item \code{municipio_codigo}: Código único do município.
\item \code{agencia_codigo}: Código da agência à qual a UC está atualmente alocada.
\item \code{dias_coleta}: Número de dias de coleta na UC.
\item \code{viagens}: Número de viagens necessárias para a coleta na UC.
}}

\item{agencias}{Um \code{tibble} ou \code{data.frame} contendo informações sobre as agências selecionáveis, incluindo:
\itemize{
\item \code{agencia_codigo}: Código único da agência.
\item \code{uc_agencia_max}: Número máximo de UCs que a agência pode atender.
\item \code{custo_fixo}: Custo fixo associado à agência.
}}

\item{custo_litro_combustivel}{Custo do combustível por litro (em R$). Padrão: 6.}

\item{custo_hora_viagem}{Custo de cada hora de viagem (em R$). Padrão: 10.}

\item{kml}{Consumo médio de combustível do veículo (em km/l). Padrão: 10.}

\item{valor_diaria}{Valor da diária para deslocamentos (em R$). Padrão: 335.}

\item{diarias_entrevistador_max}{Máximo de diárias que um entrevistador pode receber no período de referência. Padrão: \code{Inf}.}

\item{remuneracao_entrevistador}{Remuneração por entrevistador para todo o período de referência. Padrão: 0.}

\item{n_entrevistadores_min}{Número mínimo de entrevistadores por agência. Padrão: 1.}

\item{dias_coleta_entrevistador_max}{Número de dias de coleta por entrevistador.}

\item{dias_treinamento}{Número de dias/diárias para treinamento. Padrão: 0 (nenhum treinamento).}

\item{agencias_treinadas}{(Opcional) Um vetor de caracteres com os códigos das agências que já foram treinadas e não terão custo de treinamento. Padrão: NULL.}

\item{agencias_treinamento}{Código da(s) agência(s) onde o treinamento será realizado.}

\item{distancias_ucs}{Um \code{tibble} ou \code{data.frame} com as distâncias entre UCs e agências, incluindo:
\itemize{
\item \code{uc}: Código da UC.
\item \code{agencia_codigo}: Código da agência.
\item \code{distancia_km}: Distância em quilômetros entre a UC e a agência.
\item \code{duracao_horas}: Duração da viagem em horas entre a UC e a agência.
\item \code{diaria_municipio}: Indica se é necessária uma diária para deslocamento entre a UC e a agência, considerando o município da UC.
\item \code{diaria_pernoite}: Indica se é necessária uma diária com pernoite para deslocamento entre a UC e a agência.
}}

\item{distancias_agencias}{Um \code{tibble} ou \code{data.frame} com as distâncias entre as agências, incluindo:
\itemize{
\item \code{agencia_codigo_orig}: Código da agência de origem.
\item \code{agencia_codigo_dest}: Código da agência de destino.
\item \code{distancia_km}: Distância em quilômetros entre a agência de origem e a de destino.
\item \code{duracao_horas}: Duração da viagem em horas entre a agência de origem e a de destino.
}}

\item{uc_agencia_min}{Número mínimo de UCs por agência ativa. Padrão: 1.}

\item{adicional_troca_jurisdicao}{Custo adicional quando há troca de agência de coleta. Padrão: 0.}

\item{resultado_completo}{(Opcional) Um valor lógico indicando se deve ser retornado um resultado mais completo, incluindo informações sobre todas as combinações de UCs e agências. Padrão: FALSE.}

\item{solver}{Qual ferramenta para solução do modelo de otimização utilizar. Padrão: "symphony". Outras opções: "glpk", "cbc" (instalação manual).}

\item{...}{Opções adicionais para o solver.}
}
\value{
Uma lista contendo:
\itemize{
\itemize{
\item \code{resultado_municipios_jurisdicao}: Um \code{tibble} com os municipios e suas alocações originais (jurisdição), incluindo custos de deslocamento.
\item \code{resultado_agencias_jurisdicao}: Um \code{tibble} com as agências e suas alocações originais (jurisdição), incluindo custos fixos, custos de deslocamento e número de UCs alocadas.
\item \code{resultado_municipios_otimo}: Um \code{tibble} com os municípios e suas alocações otimizadas, incluindo custos de deslocamento.
\item \code{resultado_agencias_otimo}: Um \code{tibble} com as agências e suas alocações otimizadas, incluindo custos fixos, custos de deslocamento, número de UCs alocadas e número de entrevistadores.
\item \code{ucs_agencias_todas} (opcional): Um \code{tibble} com todas as combinações de UCs e agências, incluindo distâncias, custos e informações sobre diárias (retornado apenas se \code{resultado_completo} for TRUE).
\item \code{otimizacao} (opcional): O resultado completo da otimização (retornado apenas se \code{resultado_completo} for TRUE).
}
}
}
\description{
Esta função realiza a alocação otimizada de Unidades de Coleta (UCs) a agências, com o objetivo de minimizar os custos totais de deslocamento e operação. A alocação leva em consideração restrições de capacidade das agências, custos de deslocamento (combustível, tempo de viagem e diárias), custos fixos das agências e custos de treinamento.
}
