% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alocar_ucs.R
\name{alocar_ucs}
\alias{alocar_ucs}
\title{Alocação Otimizada de Unidades de Coleta (UCs) a Agências}
\usage{
alocar_ucs(
  ucs,
  agencias = NULL,
  custo_litro_combustivel = 6,
  custo_hora_viagem = 10,
  kml = 10,
  valor_diaria = 335,
  custo_fixo = 500,
  dias_treinamento = 5.5,
  dist_diaria_km = 100,
  agencias_treinadas = NULL,
  agencias_treinamento = NULL,
  distancias_ucs,
  distancias_agencias,
  min_uc_agencia = 1,
  max_uc_agencia = NULL,
  semi_centralizada = NULL,
  adicional_troca_jurisdicao = 0,
  resultado_completo = FALSE
)
}
\arguments{
\item{ucs}{Um \code{tibble} ou \code{data.frame} contendo informações sobre as UCs, incluindo:
\itemize{
\item \code{uc}: Código único da UC.
\item \code{agencia_codigo}: Código da agência à qual a UC está atualmente alocada.
\item \code{dias_coleta}: Número de dias de coleta na UC.
\item \code{viagens}: Número de viagens necessárias para a coleta na UC.
}}

\item{agencias}{Um \code{tibble} ou \code{data.frame} contendo informações sobre as agências selecionáveis, incluindo:
\itemize{
\item \code{agencia_codigo}: Código único da agência.
}}

\item{custo_litro_combustivel}{Custo do combustível por litro (em R$). Padrão: 6.}

\item{custo_hora_viagem}{Custo de cada hora de viagem (em R$). Padrão: 10.}

\item{kml}{Consumo médio de combustível do veículo (em km/l). Padrão: 10.}

\item{valor_diaria}{Valor da diária para deslocamentos (em R$). Padrão: 335.}

\item{custo_fixo}{Custo fixo mensal da agência, como salário do entrevistador, custo de monitoramento, etc. (em R$). Padrão: 500.}

\item{dias_treinamento}{Número de dias/diárias para treinamento. Padrão: 5.5}

\item{dist_diaria_km}{Distância mínima (em km) para que seja paga uma diária, no caso de UC na jurisdição da agência. No caso da UC fora da jurisdição da agência, quando a distância for menor do que \code{dist_diaria_km}, se paga meia diária, calculada como 1/2 do valor das diárias inteiras. Padrão 100}

\item{agencias_treinadas}{(Opcional) Um vetor de caracteres com os códigos das agências que já foram treinadas e não terão custo de treinamento. O custo dos APMs contratados nessas agências ainda será incluído no plano de otimização. Padrão: NULL.}

\item{agencias_treinamento}{Código da(s) agência(s) onde o treinamento será realizado.}

\item{distancias_ucs}{Um \code{tibble} ou \code{data.frame} com as distâncias entre UCs e agências, incluindo:
\itemize{
\item \code{uc}: Código da UC.
\item \code{agencia_codigo}: Código da agência.
\item \code{distancia_km}: Distância em quilômetros entre a UC e a agência.
\item \code{duracao_horas}: Duração da viagem em horas entre a UC e a agência
\item \code{diaria_municipio}: Indica se é necessária uma diária para deslocamento entre a UC e a agência, considerando o município da UC
}}

\item{distancias_agencias}{Um \code{tibble} ou \code{data.frame} com as distâncias entre as agências, incluindo:
\itemize{
\item \code{agencia_codigo_orig}: Código da agência de origem
\item \code{agencia_codigo_dest}: Código da agência de destino
\item \code{distancia_km}: Distância em quilômetros entre a agência de origem e a de destino
\item \code{duracao_horas}: Duração da viagem em horas entre a agência de origem e a de destino
}}

\item{min_uc_agencia}{Número mínimo de UCs por agência. Só válido para agências não treinadas. Padrão: 1.}

\item{max_uc_agencia}{Número máximo de UCs por agência. Padrão: NULL (ilimitado).}

\item{semi_centralizada}{(Opcional) Um vetor de caracteres com os códigos das agências que não terão limite máximo de UCs alocadas. Padrão: NULL.}

\item{adicional_troca_jurisdicao}{Custo adicional quando há troca de agência de coleta. Padrão 0}

\item{resultado_completo}{(Opcional) Um valor lógico indicando se deve ser retornado um resultado mais completo, incluindo informações sobre todas as combinações de UCs e agências. Padrão: FALSE.}
}
\value{
Uma lista contendo:
\itemize{
\item \code{resultado_ucs_otimo}: Um \code{tibble} com as UCs e suas alocações otimizadas, incluindo custos de deslocamento.
\item \code{resultado_ucs_jurisdicao}: Um \code{tibble} com as UCs e suas alocações originais (jurisdição), incluindo custos de deslocamento.
\item \code{resultado_agencias_otimo}: Um \code{tibble} com as agências e suas alocações otimizadas, incluindo custos fixos, custos de deslocamento e número de UCs alocadas.
\item \code{resultado_agencias_jurisdicao}: Um \code{tibble} com as agências e suas alocações originais (jurisdição), incluindo custos fixos, custos de deslocamento e número de UCs alocadas.
\item \code{ucs_agencias_todas} (opcional): Um \code{tibble} com todas as combinações de UCs e agências, incluindo distâncias, custos e informações sobre diárias (retornado apenas se \code{resultado_completo} for TRUE).
}
}
\description{
Esta função realiza a alocação otimizada de Unidades de Coleta (UCs) a agências, com o objetivo de minimizar os custos totais de deslocamento e operação. A alocação leva em consideração restrições de capacidade das agências, custos de deslocamento (combustível, tempo de viagem e diárias), custos fixos das agências e custos de treinamento.
}
