% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce_restricoes.R
\name{.anotar_restricoes}
\alias{.anotar_restricoes}
\title{Anotar data frame de UCs com rótulos de restrições}
\usage{
.anotar_restricoes(df, restricoes)
}
\description{
Anotar data frame de UCs com rótulos de restrições
}
\keyword{internal}
