% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce_gerar_codigo.R
\name{.codificar_vetor}
\alias{.codificar_vetor}
\title{Codifica vetor de caracteres como código R}
\usage{
.codificar_vetor(x)
}
\description{
Codifica vetor de caracteres como código R
}
\keyword{internal}
