% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_restricoes.R
\name{.descrever_restricao}
\alias{.descrever_restricao}
\title{Descrever restrição em texto legível}
\usage{
.descrever_restricao(r)
}
\description{
Descrever restrição em texto legível
}
\keyword{internal}
