% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce_gerar_codigo.R
\name{.formatar_valor}
\alias{.formatar_valor}
\title{Formata um valor R escalar como string de código}
\usage{
.formatar_valor(val)
}
\description{
Formata um valor R escalar como string de código
}
\keyword{internal}
