% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce_gerar_codigo.R
\name{.gerar_mutate_proibitivo}
\alias{.gerar_mutate_proibitivo}
\title{Gerar bloco mutate com custos proibitivos}
\usage{
.gerar_mutate_proibitivo(ucs_str, agencia_codigo, operador, custo)
}
\description{
Gerar bloco mutate com custos proibitivos
}
\keyword{internal}
