% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_uc_agencias.R
\name{map_uc_agencias}
\alias{map_uc_agencias}
\title{Mapeia Unidades de Coleta (UCs) e Agências}
\usage{
map_uc_agencias(
  data,
  f = 0.05,
  source = "google",
  maptype = "roadmap",
  zoom = "auto"
)
}
\arguments{
\item{data}{Um data frame contendo as coordenadas das UCs e agências. Deve possuir colunas chamadas 'uc_lat', 'uc_lon', 'agencia_lat' e 'agencia_lon'.}

\item{f}{Um valor numérico especificando a fração pela qual expandir a caixa delimitadora em torno das UCs e agências. O padrão é 0.05.}

\item{source}{Uma string de caracteres especificando a fonte do mapa. O padrão é "google".}

\item{maptype}{Uma string de caracteres especificando o tipo de mapa. O padrão é "roadmap".}

\item{zoom}{Um número inteiro ou a string "auto" especificando o nível de zoom. Se "auto", o nível de zoom é calculado automaticamente para ajustar todas as UCs e agências.}
}
\value{
Um objeto ggplot representando o mapa com as UCs e agências sobrepostas.
}
\description{
Esta função cria um mapa exibindo a localização das unidades de coleta (UCs) e agências.
Ela utiliza o pacote 'ggmap' para obter um mapa de uma fonte especificada (por exemplo, Google Maps)
e sobrepõe as UCs e agências nele.
}
