% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce-package.R
\docType{package}
\name{orce-package}
\alias{orce-package}
\title{orce: orce: Otimização de Redes de Coleta Estatística}
\description{
O pacote orce fornece ferramentas para otimizar a seleção de agências e a alocação de unidades de coleta para cada agência selecionada. As unidades de coleta podem ser setores censitários, Unidades Primárias de Amostragem (UPAs) (para pesquisas domiciliares), escolas (para o caso da Pesquisa Nacional de Saúde do Escolar), municípios (MUNIC), etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leoniedu.github.io/orce/}
}

}
\author{
\strong{Maintainer}: Eduardo Leoni \email{eduardo.leoni@ibge.gov.br} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
