% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce.R
\name{orce_milp}
\alias{orce_milp}
\title{Alocação Otimizada com TSP usando MILP}
\usage{
orce_milp(
  ucs,
  agencias,
  distancias_ucs,
  distancias_ucs_ucs = NULL,
  peso_tsp = 0,
  ...
)
}
\arguments{
\item{ucs, agencias, distancias_ucs, distancias_ucs_ucs, peso_tsp}{Mesmos parâmetros da função orce principal}

\item{...}{Outros parâmetros passados para orce}
}
\value{
Resultado similar à função orce principal
}
\description{
Versão alternativa usando MILP (Mixed Integer Linear Programming) ao invés de MIPModel
}
