% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce_model_milp.R
\name{orce_model_milp}
\alias{orce_model_milp}
\title{Construtor do modelo OMPR usando MILPModel (vetorizado)}
\usage{
orce_model_milp(env)
}
\arguments{
\item{env}{Ambiente contendo todos os objetos necessários ao modelo.}
}
\value{
Um objeto \code{ompr::MILPModel} com variáveis, restrições e objetivo definidos.
}
\description{
Esta função constrói o modelo equivalente a \code{orce_model_mip()} porém
utilizando a backend vetorizada \code{MILPModel}, com \code{sum_expr()} e \code{colwise()}.
}
\details{
Ela recebe um ambiente com todos os objetos preparados por \code{.orce_impl}
(por exemplo: \code{n}, \code{m}, \code{p}, \code{n_uc}, \code{N}, \code{agencias_t}, \code{ucs_i},
\code{transport_cost_i_j}, \code{diarias_i_j}, \code{dias_coleta_ijt}, \code{distancias_ucs_ucs},
\code{peso_tsp}, \code{kml}, \code{custo_litro_combustivel}, \code{remuneracao_entrevistador},
\code{n_entrevistadores_min}, \code{dias_coleta_entrevistador_max}, etc.) e retorna um
\code{ompr::MILPModel} pronto para ser resolvido.
}
