% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orce_model_mip.R
\name{orce_model_mip}
\alias{orce_model_mip}
\title{Construtor padrão do modelo OMPR do ORCE}
\usage{
orce_model_mip(env)
}
\arguments{
\item{env}{Ambiente contendo todos os objetos necessários ao modelo.}
}
\value{
Um objeto \code{ompr::MIPModel} com variáveis, restrições e objetivo definidos.
}
\description{
Esta função recebe um ambiente com todos os objetos preparados por \code{.orce_impl}
(por exemplo: \code{n}, \code{m}, \code{p}, \code{n_uc}, \code{N}, \code{agencias_t}, \code{ucs_i}, \code{transport_cost_i_j},
\code{diarias_i_j}, \code{dias_coleta_ijt}, \code{distancias_nos}, \code{peso_tsp}, \code{kml},
\code{custo_litro_combustivel}, \code{remuneracao_entrevistador}, \code{n_entrevistadores_min}, etc.)
e retorna um \code{ompr::MIPModel} pronto para ser resolvido.
}
