% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_orce_tsp.R
\name{teste_orce_tsp}
\alias{teste_orce_tsp}
\title{Teste da função orce com dados simulados incluindo TSP}
\usage{
teste_orce_tsp(
  n_agencias = 3,
  n_ucs = 20,
  n_periodos = 2,
  peso_tsp = 0.5,
  solver = "cbc",
  run_orce = orce::orce,
  ...
)
}
\arguments{
\item{n_agencias}{Número de agências a criar. Padrão: 10.}

\item{n_ucs}{Número de UCs a criar. Padrão: 30.}

\item{n_periodos}{Número de períodos de coleta. Padrão: 2.}

\item{peso_tsp}{Peso TSP para o teste. Padrão: 0.5.}

\item{solver}{Solver a usar. Padrão: "cbc".}
}
\value{
Resultado da função orce com dados simulados
}
\description{
Teste da função orce com dados simulados incluindo TSP
}
\keyword{internal}
