% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_distancias.R
\name{calcula_distancias}
\alias{calcula_distancias}
\title{Calcula Distâncias Entre Origens e Destinos}
\usage{
calcula_distancias(destinos, origens, nmax = 2000)
}
\arguments{
\item{destinos}{Um objeto \code{sf} representando os pontos de destino.}

\item{origens}{Um objeto \code{sf} representando os pontos de origem.}
}
\value{
Um data frame com as distâncias (em quilômetros) e durações (em horas) entre cada par de origem e destino.
}
\description{
Esta função calcula a distância e duração entre conjuntos de pontos de origem e destino usando o serviço OSRM.
}
