% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponto_densidade.R
\name{ponto_densidade}
\alias{ponto_densidade}
\title{Calcula o Ponto de Densidade de Unidades Espaciais}
\usage{
ponto_densidade(cnefe, geoid)
}
\arguments{
\item{cnefe}{Um objeto \code{sf} contendo a geometria das unidades espaciais e o número de
estabelecimentos em cada unidade. Deve conter uma coluna chamada \code{n} com o número
de estabelecimentos e uma coluna com o código da unidade espacial, cujo nome é
especificado no argumento \code{geoid}.}

\item{geoid}{O nome da coluna em \code{cnefe} que contém o código único da unidade espacial
(e.g., "cod_setor", "cod_municipio").}
}
\value{
Um \code{data.frame} com o código da unidade espacial (\code{geoid}) e as coordenadas
do ponto de maior densidade (latitude e longitude) para cada unidade.
}
\description{
Esta função calcula a unidade com maior densidade populacional para cada unidade espacial
(e.g., setor censitário, município) em um objeto \code{sf}, com base no número de
estabelecimentos (ou outro tipo de ponto) em cada unidade.
}
\details{
A função utiliza o pacote \code{spatstat} para calcular a densidade de pontos em cada
unidade espacial. O ponto de densidade é definido como o ponto com a maior
densidade de pontos dentro da unidade. A densidade é calculada usando um kernel
gaussiano com largura de banda \code{sigma} definida como o máximo entre 10\% da
amplitude da unidade espacial e 30 metros.
}
